Move LockClients into ConnectionManager, and guard it to prevent multiple threads from calling it at once
