+++ src/main/CMakeLists.txt
@@ -19,6 +19,7 @@ add_library_unity(
   client_context.cpp
   client_data.cpp
   client_verify.cpp
+  connection_manager.cpp
   chunk_scan_state.cpp
   config.cpp
   connection.cpp
